[OSB] Fix build on clang and gcc 12 (#86)
* Add missing #include <algorithm>

This header makes use of `std::count` which is defined in `<algorithm>`. Without the `#include` we get build errors on gcc 12.

Fixes #73.

* Add missing `typename`s to appease clang

Clang, unlike gcc, requires the use of `typename` in certain contexts
when using templates, otherwise it fails with an error.

So add the missing `typename`s throughout.

Fixes #74.