Remove GitHub pre-commit workflows
I've taken the liberty of installing the pre-commit GitHub app (https://pre-commit.ci) for the imperialCHEPI org, which replaces the need for these GitHub Actions. For some reason the auto-update workflow failed this week, so I figured we might as well change over now. We use the app for loads of other projects and haven't had issues with it.